feat: rename Infopic2 to Infopic component
