+++ packages/components/src/templates/next/components/complex/Infopic/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Infopic"
+export * from "./Infopic"
