+++ packages/components/src/templates/next/components/complex/index.ts
@@ -7,5 +7,5 @@ export { default as Iframe } from "./Iframe"
 export { default as Infobar } from "./Infobar"
 export { default as InfoCards } from "./InfoCards"
 export { default as InfoCols } from "./InfoCols"
-export { default as Infopic } from "./Infopic"
+export { Infopic } from "./Infopic"
 export { default as KeyStatistics } from "./KeyStatistics"
